dom/media/webaudio/test/test_bug845960.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Crashtest for bug 845960</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

(new AudioContext()).decodeAudioData(new ArrayBuffer(0), function() {});
ok(true, "Should not crash when the optional failure callback is not specified");

</script>
</pre>
</body>
</html>
